+++ src/server.js
@@ -7,8 +7,8 @@ const { router, get, post, put, patch, del } = require('microrouter')
 const signale = require('./utils/signale')
 const pipe = require('./utils/pipe')
 
-const auth = require('./middlewares/auth')
-const demo = require('./middlewares/demo')
+const requireAuth = require('./middlewares/requireAuth')
+const blockDemo = require('./middlewares/blockDemo')
 
 const ui = require('./routes/ui')
 const tracker = require('./routes/tracker')
@@ -55,17 +55,17 @@ module.exports = micro(
 			post('/tokens', tokens.add),
 			del('/tokens/:tokenId', tokens.del),
 
-			post('/domains', pipe(auth, demo, domains.add)),
-			get('/domains', pipe(auth, domains.all)),
-			put('/domains/:domainId', pipe(auth, demo, domains.update)),
-			del('/domains/:domainId', pipe(auth, demo, domains.del)),
+			post('/domains', pipe(requireAuth, blockDemo, domains.add)),
+			get('/domains', pipe(requireAuth, domains.all)),
+			put('/domains/:domainId', pipe(requireAuth, blockDemo, domains.update)),
+			del('/domains/:domainId', pipe(requireAuth, blockDemo, domains.del)),
 
 			post('/domains/:domainId/records', records.add),
 			patch('/domains/:domainId/records/:recordId', records.update),
 
-			get('/domains/:domainId/views', pipe(auth, views.get)),
+			get('/domains/:domainId/views', pipe(requireAuth, views.get)),
 
-			get('/domains/:domainId/referrers', pipe(auth, referrers.get)),
+			get('/domains/:domainId/referrers', pipe(requireAuth, referrers.get)),
 
 			get('/*', notFound),
 			post('/*', notFound),
